feat: use templates
