+++ spacy_llm/tasks/ner.py
@@ -1,52 +1,16 @@
-from typing import Any, Callable, Iterable, List, Optional, Tuple
+from typing import Any, Callable, Iterable, List, Literal, Optional
 
-import jinja2
 from spacy.tokens import Doc, Span
 from spacy.util import filter_spans
 
-from ..compat import Literal
-from ..registry import lowercase_normalizer, registry
-from .util import SpanExample, find_substrings
+from ..registry import registry
+from .templates import read_template
+from .util.span_task import SpanTask
 
 
 @registry.llm_tasks("spacy.NER.v1")
-class NERTask:
-    _TEMPLATE_STR = """
-From the text below, extract the following entities in the following format:
-{# whitespace #}
-{%- for label in labels -%}
-{{ label }}: <comma delimited list of strings>
-{# whitespace #}
-{%- endfor -%}
-{# whitespace #}
-{%- if examples -%}
-{# whitespace #}
-Below are some examples (only use these as a guide):
-{# whitespace #}
-{# whitespace #}
-{%- for example in examples -%}
-{# whitespace #}
-Text:
-'''
-{{ example.text }}
-'''
-{# whitespace #}
-{%- for label, substrings in example.entities.items() -%}
-{{ label }}: {{ ', '.join(substrings) }}
-{# whitespace #}
-{%- endfor -%}
-{# whitespace #}
-{# whitespace #}
-{%- endfor -%}
-{%- endif -%}
-{# whitespace #}
-Here is the text that needs labeling:
-{# whitespace #}
-Text:
-'''
-{{ text }}
-'''
-    """
+class NERTask(SpanTask):
+    _TEMPLATE_STR = read_template("ner")
 
     def __init__(
         self,
@@ -71,81 +35,19 @@ Text:
         single_match (bool): If False, allow one substring to match multiple times in
             the text. If True, returns the first hit.
         """
-        self._normalizer = normalizer if normalizer else lowercase_normalizer()
-        self._label_dict = {
-            self._normalizer(label): label for label in labels.split(",")
-        }
-        self._examples = [SpanExample(**eg) for eg in examples()] if examples else None
-        self._validate_alignment(alignment_mode)
-        self._alignment_mode = alignment_mode
-        self._case_sensitive_matching = case_sensitive_matching
-        self._single_match = single_match
-
-    def _validate_alignment(self, mode):
-        # ideally, this list should be taken from spaCy, but it's not currently exposed from doc.pyx.
-        alignment_modes = ("strict", "contract", "expand")
-        if mode not in alignment_modes:
-            raise ValueError(
-                f"Unsupported alignment mode '{mode}'. Supported modes: {', '.join(alignment_modes)}"
-            )
-
-    def generate_prompts(self, docs: Iterable[Doc]) -> Iterable[str]:
-        environment = jinja2.Environment()
-        _template = environment.from_string(self._TEMPLATE_STR)
-        for doc in docs:
-            prompt = _template.render(
-                text=doc.text,
-                labels=list(self._label_dict.values()),
-                examples=self._examples,
-            )
-            yield prompt
-
-    def _format_response(self, response: str) -> Iterable[Tuple[str, Iterable[str]]]:
-        """Parse raw string response into a structured format"""
-        output = []
-        assert self._normalizer is not None
-        for line in response.strip().split("\n"):
-            # Check if the formatting we want exists
-            # <entity label>: ent1, ent2
-            if line and ":" in line:
-                label, phrases = line.split(":", 1)
-                norm_label = self._normalizer(label)
-                if norm_label in self._label_dict:
-                    # Get the phrases / spans for each entity
-                    if phrases.strip():
-                        _phrases = [p.strip() for p in phrases.strip().split(",")]
-                        output.append((self._label_dict[norm_label], _phrases))
-        return output
+        super(NERTask, self).__init__(
+            labels=labels,
+            examples=examples,
+            normalizer=normalizer,
+            alignment_mode=alignment_mode,
+            case_sensitive_matching=case_sensitive_matching,
+            single_match=single_match,
+        )
 
     def assign_spans(
         self,
         doc: Doc,
         spans: List[Span],
-    ) -> Doc:
+    ) -> None:
         """Assign spans to the document."""
         doc.set_ents(filter_spans(spans))
-        return doc
-
-    def parse_responses(
-        self, docs: Iterable[Doc], responses: Iterable[str]
-    ) -> Iterable[Doc]:
-        for doc, prompt_response in zip(docs, responses):
-            spans = []
-            for label, phrases in self._format_response(prompt_response):
-                # For each phrase, find the substrings in the text
-                # and create a Span
-                offsets = find_substrings(
-                    doc.text,
-                    phrases,
-                    case_sensitive=self._case_sensitive_matching,
-                    single_match=self._single_match,
-                )
-                for start, end in offsets:
-                    span = doc.char_span(
-                        start, end, alignment_mode=self._alignment_mode, label=label
-                    )
-                    if span is not None:
-                        spans.append(span)
-
-            self.assign_spans(doc, spans)
-            yield doc
