+++ spacy_llm/tasks/util/__init__.py
@@ -1,4 +1,5 @@
 from .examples import SpanExample
 from .parsing import find_substrings
+from .span_task import SpanTask
 
-__all__ = ["SpanExample", "find_substrings"]
+__all__ = ["SpanExample", "find_substrings", "SpanTask"]
